Remove unnecessary fields from 'summaries'
We only need the 'featured' checkbox in prose.io. A potential problem
with this is that without the 'title' field, it will allow users to edit
the filename. We can see whether this turns out to be a problem in
practice, and if so, produce a patch for prose to allow you to disable
editing of the filename.