Remove an unintended control character
One person summary in the Pombola database had a Unicode U+0096
control character in it, which has consequently been included verbatim
in these person summaries. This was causing the parsing of this data
to fail downstream.

From the context of the text, and that 0x96 is an en-dash in
Windows-1252, it seems clear that this was intended to be an
EN DASH U+2013. We have fixed this in Pombola, so if the
build_summaries.rb script is run again this won't be reimported, but
this commit fixes the bad character directly in the Prose data in this
repository so that we don't need to re-run that script.